Removed AbstractEndpointResolver
